Style: cleanup
